dep_scan.py: don't scan sources twice
The second pass is not needed because lib->sources mapping can be built
during the first pass. It was also traversing graph starting from
includes of each source file, but now it traverses starting from the
source file itself. Because of that there is a new RTS->StDb_Tables
dependency. This is a false positive, though. It happens due to few
filename clashes like:
StDb/idl/daqFileTag.idl and StRoot/RTS/include/DB/RunLog/daqFileTag.h